ran some p =2

+++ main.py
@@ -19,7 +19,7 @@ num_train_data = int(parser["PIPELINE"]["NUM_TRAIN_DATA"])
 num_val_data = int(parser["PIPELINE"]["NUM_VAL_DATA"])
 batch_size = int(parser["PIPELINE"]["BATCH_SIZE"])
 num_exp = int(parser["PIPELINE"]["NUM_EXPERIMENTS"])
-temp_power = float(parser["TEMP"]["TEMP_POWER"])
+temp_power = int(parser["TEMP"]["TEMP_POWER"])
 dataset, val_dataset = get_data(data_set_name)
 
 # Take a subset of the dataset to ease computation
